requirements.txt
opencv-python
numpy
onnxruntime
torch
pillow